+++ packages/react-widgets/src/widgets/ScatterPlotWidget.js
@@ -23,6 +23,7 @@ import { _FeatureFlags, _hasFeatureFlag } from '@carto/react-core';
  * @param  {Function} [props.tooltipFormatter] - Function to format Y axis values.
  * @param  {boolean} [props.global] - Enable/disable the viewport filtering in the data fetching.
  * @param  {Function} [props.onError] - Function to handle error messages from the widget.
+ * @param  {Function=} [props.onStateChange] - Callback to handle state updates of widgets
  * @param  {object} [props.wrapperProps] - Extra props to pass to [WrapperWidgetUI](https://storybook-react.carto.com/?path=/docs/widgets-wrapperwidgetui--default).
  * @param  {object} [props.noDataAlertProps] - Extra props to pass to [NoDataAlert]().
  */
@@ -40,6 +41,7 @@ function ScatterPlotWidget({
   tooltipFormatter,
   global,
   onError,
+  onStateChange,
   wrapperProps,
   noDataAlertProps
 }) {
@@ -58,6 +60,7 @@ function ScatterPlotWidget({
     },
     global,
     onError,
+    onStateChange,
     attemptRemoteCalculation: _hasFeatureFlag(_FeatureFlags.REMOTE_WIDGETS)
   });